+++ odoo/songs/upgrade/update_10_0_1_49_0.py
@@ -1,0 +1,18 @@
+# -*- coding: utf-8 -*-
+# Copyright 2020 Camptocamp SA
+# License AGPL-3.0 or later (http://www.gnu.org/licenses/agpl.html)
+
+import anthem
+
+
+@anthem.log
+def reload_translation(ctx):
+    """ update translation """
+    ctx.env['ir.module.module'].with_context(overwrite=True).search(
+        [('name', '=', 'specific_report')]
+    ).update_translations()
+
+
+@anthem.log
+def post(ctx):
+    reload_translation(ctx)