+++ odoo/local-src/stock_groupbypartner/models/stock.py
@@ -1,5 +1,6 @@
 # -*- coding: utf-8 -*-
-# © 2016-2018 Jacques-Etienne Baudoux (BCIM sprl) <je@bcim.be>
+# Copyright 2016-2020 Jacques-Etienne Baudoux (BCIM) <je@bcim.be>
+# Copyright 2019-2020 Camptocamp
 # License AGPL-3.0 or later (http://www.gnu.org/licenses/agpl).
 
 import logging