+++ odoo/local-src/procurement_sale/models/sale.py
@@ -33,6 +33,10 @@ class SaleOrder(models.Model):
 class SaleOrderLine(models.Model):
     _inherit = "sale.order.line"
 
+    date_order = fields.Datetime(
+        related="order_id.date_order", readonly="True"
+    )
+
     current_product_qty_unavailable = fields.Float(
         string='Current quantity unavailable',
         digits=dp.get_precision('Product Unit of Measure'),