+++ HISTORY.rst
@@ -21,6 +21,8 @@ latest (unreleased)
 
 **Features and Improvements**
 
+* ALCYN-2451: Fix translations (invoice)
+
 **Bugfixes**
 
 **Build**